05_Chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
gunicorn